Revert "Refactor:抽象load_texture_from_url   fix:重复请求头像"
